Created setCustomerCookie and udpateLegalAge
Also created routes that uses these methods.
setCustomerCookie sets a new cookie everytime someone logs into the website.
udpateLegalAge compares the JSESSIONID to the customerID that is saved in the cookie and if they match, updates the value of legalAge to "true" in the cookie.
Return values and controllers are mainly for testing and debugging, and probably needs to get removed when the time comes.